requirements.txt
behave==1.2.6
Appium-Python-Client==2.6.0
pytest==7.1.2
Allure-Pytest==2.9.45
selenium==4.1.0
webdriver-manager==3.8.3
pytest-xdist==2.5.0
pytest-bdd==7.2.0
pytest-html==4.1.1
allure-behave==2.9.45
selenium-page-factory==2.7
browserstack-sdk
py~=1.11.0
gitdb~=4.0.11
h11~=0.14.0
pip~=24.1.2
attrs~=23.2.0
wheel~=0.43.0
cryptography~=42.0.8
filelock~=3.15.4
setuptools~=70.3.0
Mako~=1.3.5
MarkupSafe~=2.1.5
requests~=2.32.3
trio~=0.26.0
outcome~=1.3.0.post0
sniffio~=1.3.1
exceptiongroup~=1.2.2
sortedcontainers~=2.4.0
cffi~=1.16.0
idna~=3.7
PyYAML~=6.0.1
mutf8~=1.0.6
dukpy~=0.4.0
smmap~=5.0.1
PyPAC~=0.16.4
tldextract~=5.1.2
execnet~=2.1.1
urllib3~=1.26.19
six~=1.16.0
parse~=1.20.2
psutil~=6.0.0
pluggy~=1.5.0
iniconfig~=2.0.0
certifi~=2024.7.4
pycparser~=2.22
packaging~=24.1
jinja2~=3.1.4
wsproto~=1.2.0
python-dotenv~=1.0.1
tqdm~=4.66.4